requirements.txt
uqlm>=0.3.1
google-generativeai>=1.47.0
langchain-google-genai>=2.0.0
streamlit>=1.51.0
python-dotenv>=1.2.1
pandas>=2.2.3
matplotlib>=3.9.2
seaborn>=0.11.0
numpy>=1.26.4